refactor: move constants to constant file
